requirements.txt
aiogram
psycopg2-binary
sqlalchemy
python-decouple